requirements.txt
pyyaml>=6.0
scikit-learn>=1.3.0
numpy>=1.24.0
matplotlib>=3.7.0
